pyproject.toml
[tool.poetry]
name = "bitbucket-automation"
version = "0.1.0"
description = "Automation project for Bitbacket (Mend.io)"
authors = ["Przemyslaw Lukasik <przemyslaw.lukasik@vidaa.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pytest = "^8.3.2"
pytest-xdist = "^3.6.1"
pytest-html = "^4.1.1"
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^3.8.0"
chromedriver = "^2.24.1"
chromedriver-binary = "^130.0.6670.0.0"
python-dotenv = "^1.0.1"
pytest-dotenv = "^0.5.2"
playwright = "^1.46.0"
pytest-playwright = "^0.5.1"
pytest-logger = "^1.1.1"
slugify = "^0.0.1"
python-slugify = "^8.0.4"
gitpython = "^3.1.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
     \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | docs
  )/

)
'''

[tool.isort]
profile = "black"
multi_line_output = 3